💫 feat: Configure Catch-All Routes
