+++ .eslintrc.json
@@ -1,7 +1,6 @@
 {
   "extends": [
     "next/core-web-vitals",
-    "next",
-    "plugin:@next/next/recommended"
+    "prettier"
   ]
 }
