+++ app/archive/@archive/[[...filter]]/page.js
@@ -1,0 +1,28 @@
+import Link from "next/link";
+
+import NewsList from "@/components/news-list";
+import { getAvailableNewsYears, getNewsForYear } from "@/lib/news";
+
+export default function FilteredNewsPage({ params }) {
+const filter = params.filter;
+console.log("🪚 filter:", filter);
+
+const links = getAvailableNewsYears();
+
+return (
+<header id="archive-header">
+<nav>
+<ul>
+{links.map((link) => (
+<li key={link}>
+<Link href={`/archive/${link}`}>{link}</Link>
+</li>
+))}
+</ul>
+</nav>
+</header>
+);
+// const news = getNewsForYear(newsYear);
+
+// return <NewsList news={news} />;
+}
