--- app/archive/@archive/page.js
@@ -1,21 +1,0 @@
-import Link from 'next/link';
-
-import { getAvailableNewsYears } from '@/lib/news';
-
-export default function ArchivePage() {
-  const links = getAvailableNewsYears();
-
-  return (
-    <header id="archive-header">
-      <nav>
-        <ul>
-          {links.map((link) => (
-            <li key={link}>
-              <Link href={`/archive/${link}`}>{link}</Link>
-            </li>
-          ))}
-        </ul>
-      </nav>
-    </header>
-  );
-}
